Modify file/folder structure to better organize component files
