+++ frontend/src/components/Pins/Index/PinIndex.css
