+++ frontend/src/components/Pins/Show/PinShow.css
