+++ frontend/src/components/Pins/Show/PinShowOptions.tsx
