+++ frontend/src/components/Profiles/Boards/BoardShow.css
