+++ frontend/src/components/Profiles/SelectorBar.tsx
